backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
websockets==12.0
sqlalchemy==2.0.27
aiosqlite==0.19.0
pydantic==2.6.1
pydantic-settings==2.1.0
treys==0.1.8
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.9
aiofiles==23.2.1
